# Specialize `StepBy<Range<{integer}>>`